Cleanup src/coreclr/src/inc/src/metadata.h
Remove #include of private implementation headers
Fix build breaks
