+++ src/coreclr/src/ildasm/ildasmpch.h
@@ -22,4 +22,10 @@
 #include <htmlhelp.h>
 #include <conio.h>
 
+#include "mdfileformat.h"
+
+#ifndef Debug_ReportError
+#define Debug_ReportError(strMessage)
+#endif
+
 #endif
