+++ src/coreclr/src/md/enc/stdafx.h
@@ -19,6 +19,11 @@
 #include <cor.h>
 #include <corpriv.h>
 
+#include "../hotdata/hotheap.h"
+#include <metamodelro.h>
+#include <liteweightstgdb.h>
+
+
 #include "mdcommon.h"
 
 #include "utsem.h"
