+++ src/coreclr/src/md/runtime/stdafx.h
@@ -19,6 +19,10 @@
 #include <cor.h>
 #include <corpriv.h>
 
+#include "../hotdata/hotheap.h"
+#include <metamodelro.h>
+#include <liteweightstgdb.h>
+
 #include "mdcommon.h"
 
 #endif  // __STDAFX_H_
