+++ src/coreclr/src/vm/common.h
@@ -285,6 +285,7 @@ namespace Loader
 #include "log.h"
 #include "loaderheap.h"
 #include "fixuppointer.h"
+#include "stgpool.h"
 
 // src/vm
 #include "gcenv.interlocked.h"
